refactor: Move the logic of the referenced contact to a hook
