+++ react/Viewer/Footer/FooterContent.jsx
@@ -2,17 +2,11 @@ import React, { useMemo, Children, cloneElement } from 'react'
 import PropTypes from 'prop-types'
 import { makeStyles } from '@material-ui/core/styles'
 
-import { getReferencedBy, useQuery, models } from 'cozy-client'
-
 import BottomSheet, { BottomSheetHeader } from '../../BottomSheet'
 
-import { buildContactByIdsQuery } from '../queries'
 import { isValidForPanel } from '../helpers'
 import BottomSheetContent from './BottomSheetContent'
-
-const {
-  contact: { getDisplayName }
-} = models
+import useReferencedContactName from './useReferencedContactName'
 
 const useStyles = makeStyles(theme => ({
   footer: {
@@ -31,18 +25,7 @@ const FooterContent = ({ file, toolbarRef, children }) => {
 
   const toolbarProps = useMemo(() => ({ ref: toolbarRef }), [toolbarRef])
 
-  const contactIds = getReferencedBy(file, 'io.cozy.contacts').map(
-    ref => ref.id
-  )
-  const contactByIdsQuery = buildContactByIdsQuery(contactIds)
-  const { data: contactList } = useQuery(contactByIdsQuery.definition, {
-    ...contactByIdsQuery.options,
-    enabled: contactIds.length > 0
-  })
-
-  const contactsFullname = Array.isArray(contactList)
-    ? contactList.map(contact => `${getDisplayName(contact)}`).join(', ')
-    : ''
+  const { contactName, isLoadingContacts } = useReferencedContactName(file)
 
   const FooterActionButtons = Children.toArray(children).find(child => {
     return (
@@ -55,16 +38,14 @@ const FooterContent = ({ file, toolbarRef, children }) => {
     file
   })
 
-  if (
-    isValidForPanel({ file }) &&
-    (contactsFullname || contactIds.length === 0)
-  ) {
+  // We have to wait for the Contact request to finish before rendering `BottomSheet`, because it doesn't handle async well.
+  if (isValidForPanel({ file }) && !isLoadingContacts) {
     return (
       <BottomSheet toolbarProps={toolbarProps}>
         <BottomSheetHeader className="u-ph-1 u-pb-1">
           {FooterActionButtonsWithFile}
         </BottomSheetHeader>
-        <BottomSheetContent file={file} contactsFullname={contactsFullname} />
+        <BottomSheetContent file={file} contactsFullname={contactName} />
       </BottomSheet>
     )
   }
